// src/model/v1/device/Category.model.ts
export type TDeviceCategory = {
    id: string | undefined;
    createdAt: string | undefined;
    updatedAt: string | undefined;
    isDeleted: boolean | undefined;
    isActive: boolean | undefined;
    name: string | undefined;
    logo: string | undefined | null | null;
    note: string | undefined | null | null;
    deviceCategoryCode: string | undefined;
    type: string | undefined;
    vendorCode: string | undefined | null | null;
};